+++ NEWS.md
@@ -1,7 +1,7 @@
 # plot2 1.99.0.9000
 
 * Support ggplot2 4.0.0
-* Added Shiny app
+* Added Shiny app (`plot2::create_interactively()`), with file upload
 * Support `y.transform = "reverse"`
 * Support `legend.nrow`
 * Support back-to-back (= tornado, = butterfly) plots